moved outlier detector to monitoring module
