+++ src/ctapipe/monitoring/outlier.py
